long value

+++ src/main/scala/commons/money/CurrencyEnum.scala
@@ -5,8 +5,8 @@ object CurrencyEnum extends Enumeration {
   case class Currency(val name: String, val slug: String,
                       val short: String, val p: String,
                       val left: Boolean = true) extends Val(slug) {
-    def amount(a: Double): CurrencyAmount = CurrencyAmount(a, this)
-    def euroAmount(a: Double): Option[CurrencyAmount] = EURCurrencyRate.calculate(CurrencyEnum.EUR, this)(a).map(amount)
+    def amount(a: Long): CurrencyAmount = CurrencyAmount(a, this)
+    def euroAmount(a: Long): Option[CurrencyAmount] = EURCurrencyRate.calculate(CurrencyEnum.EUR, this)(a).map(amount)
     override def toString = short
   }
